requirements.txt
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
py4j==0.10.9.7
pybase64==1.3.1
PyBoof==0.43.1
pypng==0.20220715.0
pytube==15.0.0
qrcode==7.4.2
six==1.16.0
tqdm==4.66.1
transforms3d==0.4.1
typing_extensions==4.9.0
Os-sys==10.1.0
